% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_select.R
\name{forward_select}
\alias{forward_select}
\title{Select features based on feature selection}
\usage{
forward_select(train_data, test_data, min_features = 2, max_features = 10)
}
\arguments{
\item{min_features}{double. number of minimum features to select}

\item{max_features}{double. number of maximum features to select}

\item{X}{tibble. training dataset}

\item{y}{tibble. test dataset}
}
\value{
vector. The indexes of selected features
}
\description{
Select features using forward selection algorithm. 
  It starts as an empty model, and add the variable with the  
  highest improvement in the accuracy of the model. The process
  is iteratively repeated and it stops when the remaining variables 
  doesn't improve the accuracy of the model.
}
\examples{
forward_select(
  train_data, test_data, max_features=5
)
}
