% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_feature_elimination.R
\name{recursive_feature_elimination}
\alias{recursive_feature_elimination}
\title{Select features by Recursive Feature Elimination}
\usage{
recursive_feature_elimination(scorer, X, y, n_features_to_select = None)
}
\arguments{
\item{scorer}{A user-defined function that fits an estimnator and returns a score.}

\item{X}{features}

\item{y}{response}

\item{n_features_to_select}{The target number of features.}
}
\value{
A vector of feature indices representing selected features in ranked order.
}
\description{
Iteratively fit and score an estimator and eliminate features
based on the score.

The lower the score, the more favourable for the feature.
If your score is better the larger it is, then you should return the
negative of that score, for example negative MSE.
}
\examples{
# TBA
}
