% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{Feature selector that performs simmulated annealing to select features.}
\usage{
simulated_annealing(scorer, X, y, c = 1, iterations = 100, bools = FALSE)
}
\arguments{
\item{scorer}{A custom user-supplied function that accepts X and y (as defined below) as input and returns the error of the datasets.}

\item{X}{data frame of training features}

\item{y}{data frame of training targets}

\item{c}{control rate of feature perturbation}

\item{iterations}{number of iterations}

\item{bools}{If true function returns array of boolean values instead of column indicies}
}
\value{
array of selected features
}
\description{
Algorithm randomly chooses a set of features, trains on them, scores the model.
Then the algorithm slightly modifies the chosen features randomly and tests to see
if the model improves. If there is improvement, the newer model is kept, if not the 
algorithm tests to see if the worse model is still kept based on a acceptance 
probability that decreases as iterations continue and if the model performs worse.
}
\examples{

}
