% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing_select.R
\name{simulated_annealing_select}
\alias{simulated_annealing_select}
\title{Select features using simmulated annealing}
\usage{
simulated_annealing_select(
  model,
  X,
  y,
  c = 1,
  iterations = 100,
  random_state = NULL
)
}
\arguments{
\item{model}{supervised learning model}

\item{X}{data frame of training features}

\item{y}{data frame of training targets}

\item{c}{control rate of feature perturbation}

\item{iterations}{number of iterations}

\item{random_state}{random state of pseudo-random number generators}
}
\value{
vector of selected features
}
\description{
The algorithm randomly chooses a set of features, trains on them, scores the model. 
Then the algorithm slightly modifies the chosen features randomly and tests to see
if the model improves. If there is improvement, the newer model is kept, if not the 
algorithm tests to see if the worse model is still kept based on a acceptance 
probability that decreases as iterations continue and if the model performs worse.
}
