% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_threshold_select.R
\name{variance_threshold_select}
\alias{variance_threshold_select}
\title{Select features based on variance threshold}
\usage{
variance_threshold_select(data, threshold = 0)
}
\arguments{
\item{data}{tibble. The features to select from}

\item{threshold}{double. The variance threshold}
}
\value{
vector. The indexes of selected features
}
\description{
Find out about the variance of each feature and
filter out the ones below a certain threshold
}
\examples{
variance_threshold_select(
  data.frame(x1=c(1,2,3,4,5), x2=c(0,0,0,0,0), x3=c(1,1,1,1,1))
)
}
